WIP (90%) - Functional MenuView

+++ AproPOS UI/Models/MenuItemModel.swift
@@ -11,17 +11,15 @@ import UIKit // for UIImage
 
 struct MenuItemModel: Identifiable, Codable, Equatable {
     @DocumentID public var id: String?
-    //var name: String
     var price: Decimal
     var estimatedServingTime: Int // minutes
     var warnings: [String] // vegetarian, vegan, gluten, allergen, alcohol
-    var ingredients: [String: Int]
+    var ingredients: [String: Decimal]
     var image: Data // to conform to Codable
     var status: [String: [String]] // available, unavailable
     
-    init(id: String = UUID().uuidString, price: Decimal, estimatedServingTime: Int, warnings: [String] = [], ingredients: [String: Int] = [:], image: UIImage = UIImage(named: "defaultMenuItemImage")!, status: [String: [String]] = ["available": []]) {
+    init(id: String = UUID().uuidString, price: Decimal, estimatedServingTime: Int, warnings: [String] = [], ingredients: [String: Decimal] = [:], image: UIImage = UIImage(named: "defaultMenuItemImage")!, status: [String: [String]] = ["available": []]) {
         self.id = id
-        //self.name = name
         self.price = price
         self.estimatedServingTime = estimatedServingTime
         self.warnings = warnings
